% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_axes.R
\name{on_axes}
\alias{on_axes}
\title{Is point \verb{(x,y)} on an axis of the plane?}
\usage{
on_axes(x, y)
}
\arguments{
\item{x}{Numeric}

\item{y}{Numeric}
}
\value{
\code{TRUE} or \code{FALSE}, according to whether \verb{(x,y)} lies on an axis
}
\description{
The point \verb{(x,y)} lies on an axis if at least one of the coordinates is zero.
}
\examples{
on_axes(0, 0) # TRUE, this is the origin
on_axes(1, 0) # TRUE, this is on x-axis
on_axes(0, 1) # TRUE, this is on y-axis
on_axes(1, 1) # FALSE
}
