% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrant.R
\name{quadrant}
\alias{quadrant}
\title{Get quadrant of \verb{(x,y)} coordinates}
\usage{
quadrant(x, y)
}
\arguments{
\item{x}{Numeric}

\item{y}{Numeric}
}
\value{
Quadrant specified in Roman numerals, or message that point is not in
any quadrant.
}
\description{
The quadrant definitions are those found in
\href{https://en.wikipedia.org/wiki/Quadrant_(plane_geometry)}{Wikipedia}
}
\examples{
quadrant(1, 2),  # "I"
quadrant(-1, 2), # "II"
quadrant(-1,-2), # "III"
quadrant(1, -2), # "IV"
quadrant(1, 0)   # "No quadrant"
}
